pyproject.toml
[project]
name = "floss-ais"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.7",
    "numpy>=2.2.6",
    "omegaconf>=2.3.0",
    "scipy>=1.15.3",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
]
